requirements.txt
click==7.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
MarkupSafe==1.1.1
Werkzeug==1.0.1
pip==23.1.2
absl-py==1.4.0
blinker==1.6.2
build==0.10.0
click==8.1.3
cloudinary==1.33.0
docutils==0.19
flask==2.3.2
gast==0.4.0
google-auth==2.17.1
importlib-metadata==6.1.0
itsdangerous==2.1.2
jaraco-classes==3.2.3
keyring==23.13.1
markdown-it-py==2.2.0
mdurl==0.1.2
more-itertools==9.1.0
pefile==2023.2.7
pillow==9.4.0
pkginfo==1.9.6
pycocotools==2.0.6
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
pyparsing==3.0.9
pyproject-hooks==1.0.0
pytz==2023.3
pywin32-ctypes==0.2.0
readme-renderer==37.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.3.3
tensorflow==2.12.0
tomli==2.0.1
twine==4.0.2
werkzeug==2.3.4
zipp==3.15.0
altgraph==0.17.3
aniso8601==9.0.1
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
certifi==2023.5.7
cffi==1.15.1
colorama==0.4.6
comm==0.1.2
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
flask-restx==1.0.3
flit-core==3.8.0
idna==3.4
ipykernel==6.19.2
ipython==8.10.0
ipython_genutils==0.2.0
jedi==0.18.1
jinja2==3.1.2
jsonschema==4.17.3
protobuf==3.20.3
libsodium==1.0.2
lxml==4.9.2
markupsafe==2.1.1
matplotlib-inline==0.1.6
mistune==0.8.4
nest-asyncio==1.5.6
cryptography==3.4.7
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.5.2
prometheus_client==0.1

pdfkit~=1.0.0
six~=1.16.0
setuptools~=66.0.0
jupyter_core~=5.3.0
zlib~=1.2.13
